app/requirements.txt
streamlit
google-generativeai
python-dotenv
altair==4.1.0
pillow
sqlite-vec
st-pages
